xcb: fix integer underflow crash when dragging out of a window
That is, having a button press and then moving out of the window.
This is because such an event generates a negative mouse poisition.
The solution applied here was to convert the mouse position types
from u32 to i16.

+++ src/main.zig
@@ -152,8 +152,8 @@ pub const Event = struct {
             scroll_y: i2,
         },
         mouse_motion: struct {
-            x: u32,
-            y: u32,
+            x: i16,
+            y: i16,
         },
         focus_in: void,
         focus_out: void,
